app/requirements.txt
fastapi
psycopg2-binary
sqlalchemy
prometheus_client
prometheus_fastapi_instrumentator
